[FIX] memcache: If memcache was enabled but without any servers being specified (which is the default) then Tiki became unusable (because prefs are cached), requiring manually editing the database to recover - so, added a checks on critical methods, let the other options be used (such as key_prefix) and display a warning that all is not well. Also add a usable default value in the hint for memcache_servers.
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/11.x@46228 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/cache/memcachelib.php
@@ -35,9 +35,12 @@ class Memcachelib
 		// preferences are overwritten in local.php (if defined)
 		require("db/{$tikidomainslash}local.php");
 
+		$this->options  = $memcached_options;
+
 		if (!$memcached_servers || (!empty($memcached_options) && !$memcached_options['enabled']) || ! class_exists('Memcache') ) {
 			$this->memcache = FALSE;
-			$this->options  = array( 'enabled' => FALSE );
+			$this->options['enabled'] = FALSE;
+			TikiLib::lib('errorreport')->report(tra('Memcache could not be initialised, please check your settings.'));
 		} else {
 			if ( $memcached_options['compress'] == 'y' ) {
 				$memcached_options['flags'] = MEMCACHE_COMPRESSED;
@@ -46,7 +49,6 @@ class Memcachelib
 				$memcached_options['flags'] = 0;
 			}
 
-			$this->options  = $memcached_options;
 			$this->memcache = new Memcache();
 			foreach ($memcached_servers as $server) {
 				if ( $server['host'] == 'localhost' ) {
@@ -110,7 +112,11 @@ class Memcachelib
 	function get($key, $default=NULL)
 	{
 		$key = $this->buildKey($key);
-		$val = $this->memcache->get($key);
+		if ($this->memcache) {
+			$val = $this->memcache->get($key);
+		} else {
+			return $default;
+		}
 		return ($val !== NULL) ? $val : $default;
 	}
 
@@ -173,8 +179,10 @@ class Memcachelib
 	 */
 	function delete($key)
 	{
-		$key = $this->buildKey($key);
-		return $this->memcache->delete($key);
+		if ($this->memcache) {
+			$key = $this->buildKey($key);
+			return $this->memcache->delete($key);
+		}
 	}
 
 	/**
@@ -182,7 +190,9 @@ class Memcachelib
 	 */
 	function flush()
 	{
-		return $this->memcache->flush();
+		if ($this->memcache) {
+			return $this->memcache->flush();
+		}
 	}
 
 	/**
